middle-tier/app/requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
Flask==0.12.1
Flask-RESTful==0.3.6
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jwt==0.5.4
MarkupSafe==1.0
pycparser==2.18
pyjwt
pysha3
pytz==2018.5
requests==2.19.1
six==1.11.0
typing==3.5.3.0
urllib3==1.23
Werkzeug==0.14.1
web3==4.3.0
pycryptodome>=3.6.4
